worker_entite_micro_service/requirements.txt
kafka-python
pymongo
translate